refactor: split API bootstrap and remove dead pipeline exec adapter layer
